@@ -5,9 +5,18 @@ import (
 )
 
 const (
-	LoginMsgType    = "LoginMsg"
-	ResponseMsgType = "ResponseMsg"
-	RegisterMsgType = "RegisterMsg"
+	LoginMsgType            = "LoginMsg"
+	ResponseMsgType         = "ResponseMsg"
+	RegisterMsgType         = "RegisterMsg"
+	OnlineUsersMsgType      = "OnlineUsersMsg"
+	NotifyUserStatusMsgType = "NotifyUserStatusMsg"
+	SmsMsgType              = "SmsMsg"
+)
+
+const (
+	UserOnline  = "UserOnline"
+	UserOffline = "UserOffline"
+	UserBusy    = "UserBusy"
 )
 
 type Message struct {
@@ -26,8 +35,25 @@ type RegisterMsg struct {
 	UserIns model.User `json:"user"`
 }
 
-// 状态码 500 表示该用户未注册 200表示登录成功 400表示改用户已经占用
+// 状态码 500 表示该用户未注册 200表示登录成功 400表示改用户已经占用, 登录成功后可以显示当先用户列表
 type ResponseMsg struct {
 	Code  int    `json:code`
 	Error string `json:error` //返回的错误信息，如果没有则nil
+	// Users []string `users`
+}
+
+type OnlineUsersMsg struct {
+	Users []string `json:users`
+}
+
+type NotifyUserStatusMsg struct {
+	UserId string `json:userId`
+	Status string `json:status`
+}
+
+// 如果dstUserId = 11111111 则为广播
+type SmsMsg struct {
+	SrcUserId string `json:srcUserId`
+	DstUserId string `json:dstuserId`
+	Content   string `json:content`
 }
