@@ -1,7 +1,8 @@
 package message
 
 type User struct {
-	UserId   string `json:"userId"`
-	UserPwd  string `json:"userPwd"`
-	UserName string `json:"userName"`
+	UserId     string `json:"userId"`
+	UserPwd    string `json:"userPwd"`
+	UserName   string `json:"userName"`
+	UserStatus string `json:userStatus`
 }
